In ads.h: 1. In function set_Gt, the replenish flag seems to be doubled up in line 1096, it is repeated in line 1100, but then set conditional on replenish. 2. Line 1119, why is include_c here?  That only should come in to the set_Gt. See my point 3 below, should there be a replenish flag in here? 3. On line 1124 is the ‘asymp’ object (which is JxJ) just set to zero here? If the estimate_asymptote flag is off (line 448) this then just becomes zero? However, is it missing the condition where replenish is set to true? In which case Ht should correspond to delta * AjIsZero. I don’t see where the asymptote is set to one if it the flag for this is turned off.  On line 641 it is set to the corresponding prior centre of M20, but these are probably going to be zero (or whatever is being provided). 4. Why is Ht set twice, once on line 830 and another on line 834.  Should there be a replenish flag there somewhere?
In the simulation code (dlmBNSimMC.R)
5. the ‘replenish’ flag must also be set as active when setting up H (line 165 of dlmBNSimMC.R). In line 186 I took out the bit that depends on the state of the theta parameter until we figure out what is happening with the basic model.

In mode.R and ads.h there seems to be something wrong with the default priors.  If set.prior is set to FALSE then the data/model is not able to set up.

The ‘replenish’ flag really refers to the component in H being incremented by delta * (AjisZero).  However, when advertising is zero, the true replenishment rate depends on both c and that component.  The former being some fraction of the size of q.